Moved Java renderer to Example App.
Tested a Bitmap implementation for the renderer, it was far slower than
just using a byte array, as I should have expected.

The Java based renderer was moved because it has no place in the
library.  Since I'm already working in JNI for the library all included
renderers should be implemented in native code, its faster and makes
more sense.

+++ libusbtv/src/main/cpp/usbtv.cpp
@@ -162,7 +162,7 @@ JNIEXPORT jboolean JNICALL Java_com_arksine_libusbtv_UsbTvFrame_returnFrameToPoo
                                                                               jobject thisObj,
                                                                               jint poolIndex) {
 	if (usbtv != nullptr) {
-		return (jboolean) usbtv->clearFrameLock(poolIndex);
+		return (jboolean) usbtv->clearFrameLock((int)poolIndex);
 	} else {
 		return (jboolean) false;
 	}
